backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
beautifulsoup4==4.12.3
httpx==0.26.0
python-dotenv==1.0.1
pydantic==2.6.1 
bs4==0.0.1
playwright==1.41.2
openai
google-search-results